requirements.txt
pygame-2.1.2
numpy-1.22.0
noise-1.2.2